feat: implement transactional quiz submission and integrate progress tracking and point calculation services

+++ app/models/user.go
@@ -17,12 +17,12 @@ type User struct {
 	DateOfBirth           *time.Time
 	Gender                string
 	Phone                 *string
-	TotalPoints           int
+	TotalPoints           float64
 	TotalQuizzesCompleted int
 	AuthProvider          string
 	AuthProviderID        string
 	IsVerified            bool
-	IsActive              bool	
+	IsActive              bool
 	LastLoginAt           *time.Time
 }
 
@@ -30,6 +30,6 @@ func (User) AllowedFields() map[string]bool {
 	return map[string]bool{
 		"role":  true,
 		"email": true,
-		"name": true,
+		"name":  true,
 	}
 }
